project/pyproject.toml
[project]
name = "pokercard-masters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.0",
    "dotenv>=0.9.9",
    "jwt>=1.4.0",
    "google-auth>=2.0.0",
    "mysqlclient",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "pymysql>=1.1.2",
]
